rename householdIncome -> employmentHistory
