+++ src/applications/financial-status-report/pages/index.js
@@ -1,5 +1,5 @@
 import * as availableDebts from './availableDebts';
-import * as householdIncome from './householdIncome';
+import * as employmentHistory from './employmentHistory';
 import * as veteranInfo from './veteranInfo';
 
-export { availableDebts, householdIncome, veteranInfo };
+export { availableDebts, employmentHistory, veteranInfo };
